+++ src/Models/Callback.php
@@ -36,6 +36,8 @@ class Callback extends JsonSerializable
     protected $url;
     protected $users;
     protected $token;
+    protected $formsdataurl;
+    protected $userData;
 
     public function __construct(
         array $actions = [],
@@ -47,7 +49,9 @@ class Callback extends JsonSerializable
         CallbackDocStatus $status = null,
         string $url = "",
         array $users = [],
-        string $token = ""
+        string $token = "",
+        string $formsdataurl = "",
+        string $userData = ""
     ) {
         $this->actions = $actions;
         $this->changesurl = $changesurl;
@@ -59,6 +63,8 @@ class Callback extends JsonSerializable
         $this->url = $url;
         $this->users = $users;
         $this->token = $token;
+        $this->formsdataurl = $formsdataurl;
+        $this->userData = $userData;
     }
 
     /**
@@ -219,5 +225,37 @@ class Callback extends JsonSerializable
     public function setToken($token)
     {
         $this->token = $token;
+    }
+
+    /**
+     * Get the value of formsdataurl
+     */
+    public function getFormsDataUrl()
+    {
+        return $this->formsdataurl;
+    }
+
+    /**
+     * Set the value of formsdataurl
+     */
+    public function setFormsDataUrl($formsdataurl)
+    {
+        $this->formsdataurl = $formsdataurl;
+    }
+
+    /**
+     * Get the value of userData
+     */
+    public function getUserData()
+    {
+        return $this->userData;
+    }
+
+    /**
+     * Set the value of userData
+     */
+    public function setUserData($userData)
+    {
+        $this->userData = $userData;
     }
 }